bugfix, rearrange `load_metadata` methods
